refactor(core): extract theme and dynamic stylesheet engine to core.services.theme_service
