+++ src/core/services/__init__.py
@@ -16,6 +16,31 @@ from core.services.ipc_service import (
     check_single_instance,
 )
 
+from core.services.theme_service import (
+    ACCENT_COLORS,
+    CURRENT_ICON_THEME,
+    CURRENT_TRAY_ICON,
+    CATEGORY_EXTENSIONS,
+    FREEDESKTOP_MAP,
+    apply_app_theme,
+    detect_accent,
+    ensure_adaptive_icon_theme,
+    format_timestamp_relative,
+    get_app_icon,
+    get_category_for_filename,
+    get_file_icon,
+    get_themed_icon,
+    get_themed_tray_icon,
+    init_app_font,
+    make_faded_icon,
+    normalize_accent_name,
+    normalize_icon_theme_name,
+    normalize_theme_name,
+    normalize_tray_icon_name,
+    parse_size_to_bytes,
+    parse_time_to_sec,
+)
+
 __all__ = [
     "DM_CONNECTOR_PORT",
     "SignalEmitter",
@@ -26,4 +51,26 @@ __all__ = [
     "get_single_instance_key",
     "SingleInstanceServer",
     "check_single_instance",
+    "ACCENT_COLORS",
+    "CURRENT_ICON_THEME",
+    "CURRENT_TRAY_ICON",
+    "CATEGORY_EXTENSIONS",
+    "FREEDESKTOP_MAP",
+    "apply_app_theme",
+    "detect_accent",
+    "ensure_adaptive_icon_theme",
+    "format_timestamp_relative",
+    "get_app_icon",
+    "get_category_for_filename",
+    "get_file_icon",
+    "get_themed_icon",
+    "get_themed_tray_icon",
+    "init_app_font",
+    "make_faded_icon",
+    "normalize_accent_name",
+    "normalize_icon_theme_name",
+    "normalize_theme_name",
+    "normalize_tray_icon_name",
+    "parse_size_to_bytes",
+    "parse_time_to_sec",
 ]
